+++ src/file_1.h
@@ -1,5 +1,5 @@
 #include "file_1.h"
 
-float foo(int bar);
+double foo(int bar);
 
 float boo(float far);
